pyproject.toml
[project]
name = "flask-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "flask-login>=0.6.3",
  "flask>=3.1.1",
  "flask-sqlalchemy>=3.1.1",
  "requests>=2.32.4",
  "pillow>=11.2.1",
  "flask-migrate>=4.1.0",
  "flask-wtf>=1.2.2",
  "flask-babel>=4.0.0",
  "email-validator>=2.2.0",
  "flask-mail>=0.10.0",
]
